may fix 865

+++ packages/studio/core/src/capture/RecordAudio.ts
@@ -129,22 +129,21 @@ export namespace RecordAudio {
 
         recordingWorklet.onSaved = uuid => {
             project.trackUserCreatedSample(uuid)
-            editing.modify(() => {
-                fileBox.ifSome(oldFileBox => {
-                    editing.modify(() => {
-                        const newFileBox = AudioFileBox.create(boxGraph, uuid, box => {
-                            box.fileName.setValue(oldFileBox.fileName.getValue())
-                            box.startInSeconds.setValue(oldFileBox.startInSeconds.getValue())
-                            box.endInSeconds.setValue(oldFileBox.endInSeconds.getValue())
-                        })
-                        for (const pointer of [...oldFileBox.pointerHub.incoming()]) {
-                            pointer.refer(newFileBox)
-                        }
-                        for (const pointer of [...oldFileBox.transientMarkers.pointerHub.incoming()]) {
-                            pointer.refer(newFileBox.transientMarkers)
-                        }
-                        oldFileBox.delete()
+            fileBox.ifSome(oldFileBox => {
+                if (!oldFileBox.isAttached() || oldFileBox.pointerHub.isEmpty()) {return}
+                editing.modify(() => {
+                    const newFileBox = AudioFileBox.create(boxGraph, uuid, box => {
+                        box.fileName.setValue(oldFileBox.fileName.getValue())
+                        box.startInSeconds.setValue(oldFileBox.startInSeconds.getValue())
+                        box.endInSeconds.setValue(oldFileBox.endInSeconds.getValue())
                     })
+                    for (const pointer of [...oldFileBox.pointerHub.incoming()]) {
+                        pointer.refer(newFileBox)
+                    }
+                    for (const pointer of [...oldFileBox.transientMarkers.pointerHub.incoming()]) {
+                        pointer.refer(newFileBox.transientMarkers)
+                    }
+                    oldFileBox.delete()
                 })
             })
         }
